LayoutTests/fast/viewport/viewport-82.html

<head>
    <title>width=400% gives 400 pixels.</title>
    <meta name="viewport" content="width=400%">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(160, 320, 480, 320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
